pyproject.toml
[project]
name = "target-assistant-bot"
version = "0.1.1"
description = "Telegram bot that helps set goals, generate daily tasks and track progress via Google Sheets."
authors = [
    { name = "bivlked" }
]
license = { text = "MIT" }
requires-python = ">=3.10"
readme = "README.md"
keywords = ["telegram", "bot", "goals", "openai", "google-sheets"]
dependencies = [
    "python-telegram-bot>=20.8",
    "openai>=1.13",
    "APScheduler>=3.10",
    "python-dotenv>=1.0",
    "pytz>=2024.1",
    "gspread>=6.0",
    "gspread-formatting>=1.1",
    "google-auth>=2.29",
    "tenacity>=8.2",
]

[project.urls]
"Homepage" = "https://github.com/bivlked/target-assistant-bot"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
extend-select = ["E", "F"]
ignore = ["F401", "F841", "E402", "F821", "E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=. --cov-fail-under=90"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta" 
